Add 'Re-implement throttle' patch
This connection throttle takes care of pings by kicking in only if its clearly a join try.
Additionally it waits to be able to send a kick message to the client.
This implementation sends a nice german kick message containing a little highered number of the actual seconds he needs to wait. So this is not intended to be a real fix.

Whether throttle is logged can be specified by a system property, defaults to false.